requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.103.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
pydantic==1.10.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
